Separated host mapped memory types to read/write variants, bumped version to 0.7.0

+++ src/tephra/device/queue_map.cpp
@@ -145,7 +145,7 @@ void QueueMap::assignVkQueueHandles(const LogicalDevice* logicalDevice, ArrayVie
             streakStart = streakEnd;
         }
     }
-    processSharedQueues(streakStart, indices.size());
+    processSharedQueues(streakStart, static_cast<uint32_t>(indices.size()));
 }
 
 ScratchVector<std::unique_lock<Mutex>> QueueMap::lockPhysicalQueues() {
